// src/app/api/model/account-response.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AccountResponse { 
    id?: string;
    customerId?: string;
    balance?: number;
    ibanCountry?: string;
    ibanNormalized?: string;
    ibanDisplay?: string;
}