// src/app/api/model/customer-response.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AccountSummary } from './account-summary';


export interface CustomerResponse { 
    id?: string;
    firstName?: string;
    lastName?: string;
    email?: string;
    externalAuthId?: string;
    externalAuthRealm?: string;
    accounts?: Array<AccountSummary>;
}